[#593] Move connection types into separate file
